+++ lib/local-links-manager/export/link_status_exporter.rb
@@ -1,0 +1,27 @@
+require 'csv'
+
+module LocalLinksManager
+  module Export
+    class LinkStatusExporter
+      HEADINGS = %w(status count)
+
+      def self.homepage_links_status_csv
+        CSV.generate do |csv|
+          csv << HEADINGS
+          LocalAuthority.group(:status).count.each do |key, value|
+            csv << [key || "nil", value]
+          end
+        end
+      end
+
+      def self.links_status_csv
+        CSV.generate do |csv|
+          csv << HEADINGS
+          Link.enabled_links.group(:status).count.each do |key, value|
+            csv << [key || "nil", value]
+          end
+        end
+      end
+    end
+  end
+end
